+++ gulpfile.ts
@@ -12,7 +12,6 @@ const rename = require("gulp-rename");
 const mocha = require("gulp-mocha");
 const chai = require("chai");
 const tslint = require("gulp-tslint");
-const stylish = require("tslint-stylish");
 const sourcemaps = require("gulp-sourcemaps");
 const istanbul = require("gulp-istanbul");
 const remapIstanbul = require("remap-istanbul/lib/gulpRemapIstanbul");
@@ -78,6 +77,16 @@ export class Gulpfile {
     browserCopyPlatformTools() {
         return gulp.src("./src/platform/BrowserPlatformTools.template")
             .pipe(rename("PlatformTools.ts"))
+            .pipe(gulp.dest("./build/browser/src/platform"));
+    }
+
+    /**
+     * Adds dummy classes for disabled drivers (replacement is done via browser entry point in package.json)
+     */
+    @Task()
+    browserCopyDisabledDriversDummy() {
+        return gulp.src("./src/platform/BrowserDisabledDriversDummy.template")
+            .pipe(rename("BrowserDisabledDriversDummy.ts"))
             .pipe(gulp.dest("./build/browser/src/platform"));
     }
 
@@ -217,7 +226,7 @@ export class Gulpfile {
     package() {
         return [
             "clean",
-            ["browserCopySources", "browserCopyPlatformTools"],
+            ["browserCopySources", "browserCopyPlatformTools", "browserCopyDisabledDriversDummy"],
             ["packageCompile", "browserCompile"],
             "packageMoveCompiledFiles",
             [
@@ -257,8 +266,12 @@ export class Gulpfile {
     @Task()
     tslint() {
         return gulp.src(["./src/**/*.ts", "./test/**/*.ts", "./sample/**/*.ts"])
-            .pipe(tslint())
-            .pipe(tslint.report(stylish, {
+            .pipe(
+                tslint({
+                    formatter: "stylish"
+                })
+            )
+            .pipe(tslint.report({
                 emitError: true,
                 sort: true,
                 bell: true