+++ src/driver/DriverUtils.ts
@@ -1,4 +1,7 @@
-/**
+import { Driver } from "./Driver";
+import { shorten } from "../util/StringUtils";
+
+    /**
  * Common driver utility functions.
  */
 export class DriverUtils {
@@ -14,29 +17,43 @@ export class DriverUtils {
     static buildDriverOptions(options: any, buildOptions?: { useSid: boolean }): any {
         if (options.url) {
             const parsedUrl = this.parseConnectionUrl(options.url);
+            let urlDriverOptions: any = {
+                type: parsedUrl.type,
+                host: parsedUrl.host,
+                username: parsedUrl.username,
+                password: parsedUrl.password,
+                port: parsedUrl.port,
+                database: parsedUrl.database
+            };
             if (buildOptions && buildOptions.useSid) {
-                const urlDriverOptions: any = {
-                    type: options.type || parsedUrl.type,
-                    host: options.host || parsedUrl.host,
-                    username: options.username || parsedUrl.username,
-                    password: options.password || parsedUrl.password,
-                    port: options.port || parsedUrl.port,
-                    sid: options.sid || options.database || parsedUrl.database
-                };
-                return Object.assign(urlDriverOptions, options);
-            } else {
-                const urlDriverOptions: any = {
-                    type: options.type || parsedUrl.type,
-                    host: options.host || parsedUrl.host,
-                    username: options.username || parsedUrl.username,
-                    password: options.password || parsedUrl.password,
-                    port: options.port || parsedUrl.port,
-                    database: options.database || parsedUrl.database
-                };
-                return Object.assign(urlDriverOptions, options);
+                urlDriverOptions.sid = parsedUrl.database;
             }
+            return Object.assign({}, options, urlDriverOptions);
         }
         return Object.assign({}, options);
+    }
+
+    /**
+     * Builds column alias from given alias name and column name,
+     * If alias length is greater than the limit (if any) allowed by the current
+     * driver, abbreviates the longest part (alias or column name) in the resulting
+     * alias.
+     *
+     * @param driver Current `Driver`.
+     * @param alias Alias part.
+     * @param column Name of the column to be concatened to `alias`.
+     *
+     * @return An alias allowing to select/transform the target `column`.
+     */
+    static buildColumnAlias({ maxAliasLength }: Driver, alias: string, column: string): string {
+        const columnAliasName = alias + "_" + column;
+
+        if (maxAliasLength && maxAliasLength > 0 && columnAliasName.length > maxAliasLength)
+            return alias.length > column.length
+                ? `${shorten(alias)}_${column}`
+                : `${alias}_${shorten(column)}`;
+
+        return columnAliasName;
     }
 
     // -------------------------------------------------------------------------
@@ -76,5 +93,4 @@ export class DriverUtils {
             database: afterBase || undefined
         };
     }
-
 }