+++ src/persistence/SubjectExecutor.ts
@@ -103,6 +103,9 @@ export class SubjectExecutor {
         // recompute only in the case if any listener or subscriber was really executed
         if (broadcasterResult && broadcasterResult.count > 0) {
             // console.time(".recompute");
+            this.insertSubjects.forEach(subject => subject.recompute());
+            this.updateSubjects.forEach(subject => subject.recompute());
+            this.removeSubjects.forEach(subject => subject.recompute());
             this.recompute();
             // console.timeEnd(".recompute");
         }
@@ -492,7 +495,10 @@ export class SubjectExecutor {
 
             // mongo _id remove
             if (this.queryRunner instanceof MongoQueryRunner) {
-                if (subject.metadata.objectIdColumn && subject.metadata.objectIdColumn.databaseName) {
+                if (subject.metadata.objectIdColumn
+                    && subject.metadata.objectIdColumn.databaseName
+                    && subject.metadata.objectIdColumn.databaseName !== subject.metadata.objectIdColumn.propertyName
+                ) {
                     delete subject.entity[subject.metadata.objectIdColumn.databaseName];
                 }
             }